+++ .env
@@ -1,0 +1,7 @@
+SECRET_KEY=please_change_me
+DJANGO_ALLOWED_HOSTS=localhost 0.0.0.0 127.0.0.1 [::1]
+POSTGRES_ENGINE=django.db.backends.postgresql
+POSTGRES_DB=myshowroom_db
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=123
+POSTGRES_PORT=5432
