+++ car/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from car.models import Car,CarManufacturer
+
+admin.site.register(Car)
+admin.site.register(CarManufacturer)
