+++ customer/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from customer.models import Customer,CustomerHistory
+
+
+admin.site.register(Customer)
+admin.site.register(CustomerHistory)
