+++ customer/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CustomerConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'customer'
