+++ customer/serializers.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+from customer.models import Customer,User
+
+
+class CustomerSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Customer
+        fields = '__all__'
+        read_only_fields = ('balance',)
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = '__all__'
