+++ customer/urls.py
@@ -1,0 +1,10 @@
+from rest_framework import routers
+from customer.views import CustomerViewSet, UserViewSet
+
+
+router = routers.DefaultRouter()
+router.register(r'customer', CustomerViewSet)
+router.register(r'user', UserViewSet)
+
+
+urlpatterns = router.urls
