+++ customer/views.py
@@ -1,0 +1,21 @@
+from rest_framework import viewsets, mixins
+from customer.models import Customer
+from customer.serializers import CustomerSerializer, UserSerializer
+
+
+class CustomerViewSet(mixins.ListModelMixin,
+                    mixins.CreateModelMixin,
+                    mixins.RetrieveModelMixin,
+                    mixins.UpdateModelMixin,
+                    viewsets.GenericViewSet):
+    queryset = Customer.objects.all()
+    serializer_class = CustomerSerializer
+
+
+class UserViewSet(mixins.ListModelMixin,
+                    mixins.CreateModelMixin,
+                    mixins.RetrieveModelMixin,
+                    mixins.UpdateModelMixin,
+                    viewsets.GenericViewSet):
+    serializer_class = UserSerializer
+    queryset = Customer.objects.all()
