+++ docker-compose.yml
@@ -11,6 +11,8 @@ services:
 
     depends_on:
       - db
+    env_file:
+      - .env
 
   db:
       container_name: db
