+++ provider/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from provider.models import Provider,ProviderDiscount
+
+
+admin.site.register(Provider)
+admin.site.register(ProviderDiscount)
