+++ showroom/admin.py
@@ -1,12 +1,7 @@
 from django.contrib import admin
-from .models import *
+from showroom.models import Showroom,ShowroomDiscount,ShowroomHistory
 
-admin.site.register(Car)
-admin.site.register(CarManufacturer)
-admin.site.register(Customer)
-admin.site.register(CustomerHistory)
+
 admin.site.register(Showroom)
 admin.site.register(ShowroomDiscount)
 admin.site.register(ShowroomHistory)
-admin.site.register(Provider)
-admin.site.register(ProviderDiscount)
