+++ showroom/views.py
@@ -1,28 +1,9 @@
 from django.shortcuts import render
 from rest_framework import viewsets,mixins
-
-from showroom.models import *
-from showroom.seralizers import *
-
+from showroom.models import Showroom,ShowroomDiscount
+from showroom.serializers import ShowroomSerializer,ShowroomDiscountSerializer
 
 
-class CarManufacturerViewSet(mixins.ListModelMixin,
-                    mixins.CreateModelMixin,
-                    mixins.UpdateModelMixin,
-                    mixins.RetrieveModelMixin,
-                    viewsets.GenericViewSet):
-    queryset = CarManufacturer.objects.all()
-    serializer_class = CarManufacturerSerializer
-
-
-
-class CarViewSet(mixins.ListModelMixin,
-                   mixins.CreateModelMixin,
-                   mixins.UpdateModelMixin,
-                   mixins.RetrieveModelMixin,
-                   viewsets.GenericViewSet):
-    queryset = Car.objects.all()
-    serializer_class = CarSerializer
 
 
 class ShowroomViewSet(mixins.ListModelMixin,
@@ -43,38 +24,6 @@ class ShowroomDiscountViewSet(mixins.ListModelMixin,
     serializer_class = ShowroomDiscountSerializer
 
 
-class CustomerViewSet(mixins.ListModelMixin,
-                    mixins.CreateModelMixin,
-                    mixins.RetrieveModelMixin,
-                    mixins.UpdateModelMixin,
-                    viewsets.GenericViewSet):
-    queryset = Customer.objects.all()
-    serializer_class = CustomerSerializer
 
 
-class UserViewSet(mixins.ListModelMixin,
-                    mixins.CreateModelMixin,
-                    mixins.RetrieveModelMixin,
-                    mixins.UpdateModelMixin,
-                    viewsets.GenericViewSet):
-    serializer_class = UserSerializer
-    queryset = Customer.objects.all()
-
-
-class ProviderViewSet(mixins.ListModelMixin,
-                    mixins.CreateModelMixin,
-                    mixins.RetrieveModelMixin,
-                    mixins.UpdateModelMixin,
-                    viewsets.GenericViewSet):
-    queryset = Provider.objects.all()
-    serializer_class = ProviderSerializer
-
-
-class ProviderDiscountViewSet(mixins.ListModelMixin,
-                    mixins.CreateModelMixin,
-                    mixins.RetrieveModelMixin,
-                    mixins.UpdateModelMixin,
-                    viewsets.GenericViewSet):
-    queryset = ProviderDiscount.objects.all()
-    serializer_class = ProviderDiscountSerializer
